Merge new timetable exporter

+++ rrrr_types.h
@@ -198,8 +198,8 @@ struct router_request {
 #define RTIME_THREE_DAYS  (SEC_TO_RTIME(SEC_IN_THREE_DAYS))
 
 #define UNREACHED UINT16_MAX
-#define NONE      (UINT32_MAX)
-#define WALK      (UINT32_MAX - 1)
+#define NONE      (UINT16_MAX)
+#define WALK      (UINT16_MAX - 1)
 
 #define STOP_NONE ((spidx_t) -1)
 #define ONBOARD   ((spidx_t) -2)
